Add a split panel layout component
