+++ packages/pipeline-editor/src/PropertiesPanel/index.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useRef, useState } from "react";
+import React, { useEffect, useRef } from "react";
 
 import { CommonProperties } from "@elyra/canvas";
 import { nanoid } from "nanoid";
@@ -74,104 +74,33 @@ function PropertiesPanel({
   canvasController,
   onPropertiesChange,
 }: any) {
-  const [width, setWidth] = useState(500);
-
-  const dragging = useRef(false);
-
-  useEffect(() => {
-    function handleMouseMove(e: MouseEvent) {
-      if (dragging.current) {
-        e.preventDefault();
-        const min = 300;
-        const { clientX } = e;
-        const rawPanelWidth = document.body.clientWidth - clientX;
-        const panelWidth = Math.max(
-          Math.min(document.body.clientWidth - min, rawPanelWidth),
-          min
-        );
-        setWidth(panelWidth);
-      }
-    }
-
-    function handleMouseUp() {
-      dragging.current = false;
-    }
-
-    document.addEventListener("mousemove", handleMouseMove);
-    document.addEventListener("mouseup", handleMouseUp);
-
-    return () => {
-      document.removeEventListener("mousemove", handleMouseMove);
-      document.removeEventListener("mouseup", handleMouseUp);
-    };
-  }, []);
-
-  const handleMouseDown = useCallback(() => {
-    dragging.current = true;
-  }, []);
-
   return (
     <React.Fragment>
-      <div
-        style={{
-          zIndex: 10000,
-          position: "absolute",
-          top: 0,
-          bottom: 0,
-          background: "var(--elyra-color-panel-border)",
-          width: "1px",
-          right: `${width}px`,
-        }}
-      />
-      <div
-        style={{
-          zIndex: 10000,
-          position: "absolute",
-          top: 0,
-          bottom: 0,
-          background: "var(--elyra-color-panel-bg)",
-          width: `${width}px`,
-          right: 0,
-        }}
-      >
-        <div>
-          <div style={{ display: "flex" }}>
-            <div>NODE PROPERTIES</div>
-            <div>PALETTE</div>
-          </div>
-          <div>
-            <div className="codicon codicon-close"></div>
-          </div>
+      <div>
+        <div style={{ display: "flex" }}>
+          <div>NODE PROPERTIES</div>
+          <div>PALETTE</div>
         </div>
-        <div
-          style={{
-            position: "absolute",
-            top: "35px",
-            bottom: 0,
-            overflow: "scroll",
-            width: "100%",
-          }}
-        >
-          <PropertiesContent
-            selectedNodes={selectedNodes}
-            nodes={nodes}
-            canvasController={canvasController}
-            onChange={onPropertiesChange}
-          />
+        <div>
+          <div className="codicon codicon-close"></div>
         </div>
       </div>
       <div
         style={{
-          zIndex: 10000,
           position: "absolute",
-          cursor: "col-resize",
-          top: 0,
+          top: "35px",
           bottom: 0,
-          width: "8px",
-          right: `${width - 4}px`,
+          overflow: "scroll",
+          width: "100%",
         }}
-        onMouseDown={handleMouseDown}
-      />
+      >
+        <PropertiesContent
+          selectedNodes={selectedNodes}
+          nodes={nodes}
+          canvasController={canvasController}
+          onChange={onPropertiesChange}
+        />
+      </div>
     </React.Fragment>
   );
 }
